setting year for address data
All address data for Cuyahoga County is from a 2006 project. It is sometimes dated 2010 as the year of publication, it is sometimes attributed to Cleveland (https://github.com/openaddresses/openaddresses/blob/master/sources/us-oh-cleveland.json), but the actual verified address data is 2006, and it is for the extent of Cuyahoga County.